Correções na view

+++ src/webService/WebServiceAccess.java
@@ -12,18 +12,18 @@ public class WebServiceAccess {
 	/* Inicializa a base de bandas */
 	static {
 		
-		int id = 1;
+		Long id = 1L;
 		
-		listaBandas.add( new Banda( id++, "Led Zeppelin", 1968) );
-		listaBandas.add( new Banda( id++, "AC/DC", 1973) );
-		listaBandas.add( new Banda( id++, "Rolling Stones", 1962) );
-		listaBandas.add( new Banda( id++, "Black Sabbath", 1968 ) );
-		listaBandas.add( new Banda( id++, "Cream", 1966) );
-		listaBandas.add( new Banda( id++, "The Animals", 1964) );
-		listaBandas.add( new Banda( id++, "The Who", 1964) );
-		listaBandas.add( new Banda( id++, "Creedence Clearwater Revival", 1967) );
-		listaBandas.add( new Banda( id++, "The Mamas & the Papas", 1966) );
-		listaBandas.add( new Banda( id++, "Deep Purple", 1968) );
+		listaBandas.add( new Banda( id++, "Led Zeppelin", 1968L) );
+		listaBandas.add( new Banda( id++, "AC/DC", 1973L) );
+		listaBandas.add( new Banda( id++, "Rolling Stones", 1962L) );
+		listaBandas.add( new Banda( id++, "Black Sabbath", 1968L) );
+		listaBandas.add( new Banda( id++, "Cream", 1966L) );
+		listaBandas.add( new Banda( id++, "The Animals", 1964L) );
+		listaBandas.add( new Banda( id++, "The Who", 1964L) );
+		listaBandas.add( new Banda( id++, "Creedence Clearwater Revival", 1967L) );
+		listaBandas.add( new Banda( id++, "The Mamas & the Papas", 1966L) );
+		listaBandas.add( new Banda( id++, "Deep Purple", 1968L) );
 	}
